+++ src/components/UI/Global/GlobalBar/GlobalBar.tsx
@@ -2,19 +2,17 @@ import styled from 'styled-components';
 import ExampleProfile from '../../../../assets/examples/profile-example.jpeg'
 import ProfileImage from '../../Common/Profile/Image/Image'
 import {Grid3X3, Plug, SettingGear, Update} from "../../Icons/Icons";
-import {randomColor} from "../../../../utilities/utilities";
 import {screenWidth, sizes} from "../../screenSizes";
-import {LogoIcon, LogoImage, LogoText} from "../../Logo/Logo";
-import AppsDropdown from "../../Dropdowns/AppsDropdown/AppsDropdown";
-import ProfileDropdown, {ProfileData} from "../../Dropdowns/ProfileDropdown/ProfileDropdown";
-import {useState} from "react";
+import {LogoIcon} from "../../Logo/Logo";
+import ProfileDropdown, {Profiles} from "../../Dropdowns/ProfileDropdown/ProfileDropdown";
 import useAppState from "../../../../store/appState/useAppState";
-import {setIsSettingsOpen} from "../../../../store/appState/appStateActions";
+import {TippyDropdown} from "../../Tippy/Tippy";
+import AppsDropdown from "../../Dropdowns/AppsDropdown/AppsDropdown";
 
 const StyledGlobalBar = styled.div`
   height: calc(50px + env(safe-area-inset-top));
   width: 100%;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+  border-bottom: 1px solid ${props => props.theme.ui.border.color};
   display: flex;
   align-items: center;
   background-color: ${props => props.theme.ui.backgroundColor};
@@ -44,11 +42,20 @@ const StyledBarSectionOptional = styled.div`
   }
 `
 
-const StyledGlobalButtonWrapper = styled.div`
+const StyledGlobalButton = styled.button`
+  height: 48px;
+  width: 48px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: none;
+  background-color: transparent;
+  cursor: pointer;
+  padding: 10px;
   position: relative;
 `
 
-const StyledGlobalButton = styled.button`
+const StyledGlobalItem = styled.div`
   height: 48px;
   width: 48px;
   display: flex;
@@ -67,23 +74,26 @@ const GlobalBar = () => {
         setIsDrawerOpen,
         isDrawerOpen,
         windowInnerSize: {width},
-        setIsSettingsOpen
+        setIsModalOpen
     } = useAppState()
-    const [currentDropdown, setDropdown] = useState<'apps' | 'profile' | ''>('')
 
-    const exampleProfile: ProfileData = {
-        imageSrc: ExampleProfile,
-        keyid: '75DDC3C4A499F1A1',
-        name: 'Jessica K'
-    }
-
-
-    const dropdownAppear = (dropdown: 'apps' | 'profile' | '') => {
-        if (currentDropdown === dropdown) {
-            return setDropdown('')
+    const exampleProfiles: Profiles = [
+        {
+            imageSrc: ExampleProfile,
+            keyid: '75DDC3C4A499F1A1',
+            name: 'Jessica K'
+        },
+        {
+            imageSrc: 'https://source.unsplash.com/random/200x200/?city',
+            keyid: '85CCD3D535DA1DS',
+            name: 'Private Account'
+        },
+        {
+            imageSrc: 'https://source.unsplash.com/random/200x200/?cute',
+            keyid: '96BDA5D6S2C1SDB',
+            name: 'Design Studio'
         }
-        return setDropdown(dropdown)
-    }
+    ]
 
 
     return (
@@ -100,42 +110,43 @@ const GlobalBar = () => {
                 </StyledGlobalButton>
 
                 <StyledBarSectionOptional>
-                    <StyledGlobalButtonWrapper>
-                        <StyledGlobalButton onClick={() => setIsSettingsOpen(true)}>
-                            <SettingGear size={18}/>
-                        </StyledGlobalButton>
-                    </StyledGlobalButtonWrapper>
+                    <StyledGlobalButton onClick={() => setIsModalOpen('settings')}>
+                        <SettingGear size={18}/>
+                    </StyledGlobalButton>
 
-                    <StyledGlobalButtonWrapper>
-                        <StyledGlobalButton onClick={() => dropdownAppear('apps')}>
+
+                    <TippyDropdown
+                        content={<AppsDropdown/>}
+                        interactive={true}
+                        verticalOffset={-5}
+                        horizontalOffset={-110}
+                    >
+                        <StyledGlobalItem>
                             <Grid3X3 size={18}/>
-                        </StyledGlobalButton>
-                        {
-                            currentDropdown === 'apps' && (
-                                <AppsDropdown/>
-                            )
-                        }
-                    </StyledGlobalButtonWrapper>
+                        </StyledGlobalItem>
+                    </TippyDropdown>
+
                 </StyledBarSectionOptional>
 
-                <StyledGlobalButtonWrapper>
-                    <StyledGlobalButton onClick={() => dropdownAppear('profile')}>
+
+                <TippyDropdown
+                    content={<ProfileDropdown profiles={exampleProfiles}/>}
+                    interactive={true}
+                    verticalOffset={-5}
+                    horizontalOffset={-110}
+                >
+                    <StyledGlobalItem>
                         <ProfileImage src={ExampleProfile} size='sm'/>
-                    </StyledGlobalButton>
-                    {
-                        currentDropdown === 'profile' && (
-                            <ProfileDropdown closeAction={() => dropdownAppear('')} profileData={exampleProfile}/>
-                        )
-                    }
-                </StyledGlobalButtonWrapper>
+                    </StyledGlobalItem>
+                </TippyDropdown>
+
+
                 {
                     hasUpdateAvailable && (
-                        <StyledGlobalButtonWrapper>
-                            <StyledGlobalButton onClick={() => {
-                            }}>
-                                <Update size='sm'/>
-                            </StyledGlobalButton>
-                        </StyledGlobalButtonWrapper>
+                        <StyledGlobalButton onClick={() => {
+                        }}>
+                            <Update size='sm'/>
+                        </StyledGlobalButton>
                     )
                 }
             </StyledBarSection>